requirements.txt
diskcache==5.4.0
PyGithub==1.55
python-Levenshtein==0.12.2
thefuzz==0.19.0
